+++ src/redux/favoritesSlice/favoritesOperations.js
@@ -1,6 +1,7 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
 import axios from 'axios';
 import { toast } from 'react-toastify';
+import { selectIsLoggedIn } from '../authSlice/authSelectors.js';
 
 export const fetchProductsInFavorites = createAsyncThunk(
   'favorites/fetchProductsInFavorites',
@@ -15,10 +16,40 @@ export const fetchProductsInFavorites = createAsyncThunk(
   }
 );
 
+export const fetchFavoritesFromLocal = createAsyncThunk(
+  'favorites/fetchFavoritesFromLocal',
+  async (_, thunkAPI) => {
+    try {
+      const state = thunkAPI.getState();
+      const localFavoritesIds = state.favorites?.favoritesIds || [];
+      if (!localFavoritesIds.length) {
+        return [];
+      }
+      const response = await axios.post(
+        '/favorites/not-authorized',
+        { productIds: localFavoritesIds },
+        {
+          headers: {
+            'Content-Type': 'application/json',
+          },
+        }
+      );
+      return response.data;
+    } catch (error) {
+      toast.error('Помилка при завантаженні улюблених товарів');
+      return thunkAPI.rejectWithValue(error.message);
+    }
+  }
+);
+
+
 export const addProductToFavorites = createAsyncThunk(
   'favorites/addProductToFavorites',
   async (productId, thunkAPI) => {
+    const state = thunkAPI.getState();
+    const isLoggedIn = selectIsLoggedIn(state);
     try {
+      if (!isLoggedIn) return thunkAPI.rejectWithValue('Потрібно авторизуватися');
       const response = await axios.post('/favorites', { productId });
       return response.data.products;
     } catch (error) {
@@ -31,7 +62,10 @@ export const addProductToFavorites = createAsyncThunk(
 export const removeProductFromFavorites = createAsyncThunk(
   'favorites/removeProductFromFavorites',
   async (productId, thunkAPI) => {
+    const state = thunkAPI.getState();
+    const isLoggedIn = selectIsLoggedIn(state);
     try {
+      if (!isLoggedIn) return thunkAPI.rejectWithValue('Потрібно авторизуватися');
       const response = await axios.delete('/favorites', { data: { productId } });
       return response.data;
     } catch (error) {
@@ -42,7 +76,10 @@ export const removeProductFromFavorites = createAsyncThunk(
 );
 
 export const clearFavorites = createAsyncThunk('favorites/clearFavorites', async (_, thunkAPI) => {
+  const state = thunkAPI.getState();
+  const isLoggedIn = selectIsLoggedIn(state);
   try {
+    if (!isLoggedIn) return thunkAPI.rejectWithValue('Потрібно авторизуватися');
     const response = await axios.delete('/favorites');
     return response.data;
   } catch (error) {
